requirements.txt
# Создайте файл requirements.txt со следующим содержимым:
django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
openai==1.12.0
djangorestframework-simplejwt==5.3.0
django-ipware==5.0.0
psycopg2-binary==2.9.9